wetter/pyproject.toml
[tool.poetry]
name = "wetter"
version = "0.3.1"
description = "CLI tool to parse weather data"
authors = ["Uğur Çayoğlu <cayoglu@me.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ucyo/wetter-py"
repository = "https://github.com/ucyo/wetter-py"
documentation = "https://github.com/ucyo/wetter-py"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.scripts]
wetter = "wetter.app:main"

[tool.poetry.dependencies]
python = "^3.8.1"
pandas = "^1.5.3"
requests = "^2.28.2"
platformdirs = "^3.2.0"
toml = "^0.10.2"
numexpr = "^2.8.4"
Bottleneck = "^1.3.7"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
pytest-cov = "^4.0.0"
isort = "^5.12.0"

[tool.black]
line-length = 120

[tool.pytest.ini_options]
markers = [
    "web: mark tests as web (deselect with '-m \"not web\"')",
    "long: mark tests which take a long time (deselect with '-m \"not long\"')",
]

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
